UI改善：文字サイズ調整・通貨切替・小数点対応

+++ app/src/main/java/com/example/subskill/date/AppSettings.kt
@@ -8,7 +8,7 @@ data class AppSettings(
     @PrimaryKey
     val packageName: String,
     val serviceName: String? = null,
-    val monthlyFee: Int? = null,
+    val monthlyFee: Double? = null,
     val isCandidate: Boolean = false,
     val isManualSubscription: Boolean = false
 )
